refactor github transport retry helpers
