+++ src/github.test.ts
@@ -1,6 +1,6 @@
 import test from "node:test";
 import assert from "node:assert/strict";
-import { GitHubClient, inferCopilotReviewLifecycle, isTransientGitHubCommandFailure } from "./github";
+import { GitHubClient, inferCopilotReviewLifecycle } from "./github";
 import { SupervisorConfig } from "./types";
 
 function createConfig(overrides: Partial<SupervisorConfig> = {}): SupervisorConfig {
@@ -811,222 +811,4 @@ test("GitHubClient fetches the newest unresolved review thread comments", async 
   assert.equal(threads.length, 1);
   assert.equal(threads[0]?.comments.nodes.at(-1)?.id, "comment-100");
   assert.equal(threads[0]?.comments.nodes.at(-1)?.author?.typeName, "Bot");
-});
-
-test("isTransientGitHubCommandFailure matches connection reset GraphQL failures", () => {
-  assert.equal(
-    isTransientGitHubCommandFailure(
-      'Command failed: gh pr list --repo owner/repo\nexitCode=1\nPost "https://api.github.com/graphql": read tcp 127.0.0.1:12345->140.82.112.6:443: read: connection reset by peer',
-    ),
-    true,
-  );
-  assert.equal(
-    isTransientGitHubCommandFailure("Command failed: gh pr create --repo owner/repo\nexitCode=1\npull request create failed: No commits between main and branch"),
-    false,
-  );
-});
-
-test("GitHubClient retries transient gh failures and succeeds on a later attempt", async () => {
-  const config = createConfig();
-  const calls: Array<{ command: string; args: string[] }> = [];
-  let delayCalls = 0;
-  const client = new GitHubClient(
-    config,
-    async (command, args) => {
-      calls.push({ command, args });
-      if (args[0] === "pr" && args[1] === "list" && calls.filter((call) => call.args[0] === "pr" && call.args[1] === "list").length < 3) {
-        throw new Error(
-          'Command failed: gh pr list --repo owner/repo\nexitCode=1\nPost "https://api.github.com/graphql": read tcp 127.0.0.1:12345->140.82.112.6:443: read: connection reset by peer',
-        );
-      }
-
-      if (args[0] === "api" && args[1] === "graphql") {
-        return {
-          exitCode: 0,
-          stdout: '{"data":{"repository":{"pullRequest":{"reviewRequests":{"nodes":[]},"reviews":{"nodes":[]},"timelineItems":{"nodes":[]}}}}}',
-          stderr: "",
-        };
-      }
-
-      return {
-        exitCode: 0,
-        stdout:
-          '[{"number":17,"title":"Retry gh","url":"https://example.test/pr/17","state":"OPEN","createdAt":"2026-03-13T00:00:00Z","updatedAt":"2026-03-13T00:00:00Z","isDraft":false,"reviewDecision":null,"mergeStateStatus":"CLEAN","mergeable":"MERGEABLE","headRefName":"codex/issue-105","headRefOid":"deadbeef","mergedAt":null}]',
-        stderr: "",
-      };
-    },
-    async () => {
-      delayCalls += 1;
-    },
-  );
-
-  const pr = await client.findOpenPullRequest("codex/issue-105");
-
-  assert.equal(pr?.number, 17);
-  assert.equal(calls.filter((call) => call.args[0] === "pr" && call.args[1] === "list").length, 3);
-  assert.equal(calls.filter((call) => call.args[0] === "api" && call.args[1] === "graphql").length, 1);
-  assert.equal(delayCalls, 2);
-});
-
-test("GitHubClient retry warnings redact raw gh arguments", async () => {
-  const config = createConfig();
-  const warnings: string[] = [];
-  const originalWarn = console.warn;
-  console.warn = (message?: unknown, ...args: unknown[]) => {
-    warnings.push([message, ...args].map((value) => String(value)).join(" "));
-  };
-
-  try {
-    let calls = 0;
-    const client = new GitHubClient(
-      config,
-      async (_command, args) => {
-        calls += 1;
-        if (args[0] === "api" && args[1] === "graphql" && calls === 1) {
-          throw new Error(
-            'Command failed: gh api graphql -f query=query { viewer { login secretField } }\nexitCode=1\nPost "https://api.github.com/graphql": read tcp 127.0.0.1:12345->140.82.112.6:443: read: connection reset by peer',
-          );
-        }
-
-        return {
-          exitCode: 0,
-          stdout: JSON.stringify({
-            data: {
-              repository: {
-                issue: {
-                  closedByPullRequestsReferences: {
-                    nodes: [],
-                  },
-                },
-              },
-            },
-          }),
-          stderr: "",
-        };
-      },
-      async () => undefined,
-    );
-
-    await client.getMergedPullRequestsClosingIssue(105);
-
-    assert.equal(warnings.length, 1);
-    assert.match(warnings[0] ?? "", /Transient GitHub CLI failure for gh api graphql/);
-    assert.match(warnings[0] ?? "", /\+\d+ arg/);
-    assert.doesNotMatch(warnings[0] ?? "", /secretField/);
-    assert.doesNotMatch(warnings[0] ?? "", /query=query/);
-  } finally {
-    console.warn = originalWarn;
-  }
-});
-
-test("GitHubClient terminal transient failure redacts raw gh arguments", async () => {
-  const config = createConfig();
-  const client = new GitHubClient(
-    config,
-    async () => {
-      throw new Error(
-        'Command failed: gh api graphql -f query=query { viewer { login secretField } }\nexitCode=1\nPost "https://api.github.com/graphql": read tcp 127.0.0.1:12345->140.82.112.6:443: read: connection reset by peer',
-      );
-    },
-    async () => undefined,
-  );
-
-  await assert.rejects(
-    () => client.getMergedPullRequestsClosingIssue(44),
-    (error: unknown) => {
-      assert.ok(error instanceof Error);
-      assert.match(error.message, /Transient GitHub CLI failure after 3 attempts: gh api graphql/);
-      assert.match(error.message, /\+\d+ arg/);
-      assert.doesNotMatch(error.message, /secretField/);
-      assert.doesNotMatch(error.message, /query=query/);
-      return true;
-    },
-  );
-});
-
-test("GitHubClient does not retry non-transient gh failures", async () => {
-  const config = createConfig();
-  let calls = 0;
-  const client = new GitHubClient(
-    config,
-    async () => {
-      calls += 1;
-      throw new Error(
-        "Command failed: gh pr create --repo owner/repo\nexitCode=1\npull request create failed: No commits between main and codex/issue-105",
-      );
-    },
-    async () => undefined,
-  );
-
-  await assert.rejects(
-    () =>
-      client.createPullRequest(
-        {
-          number: 105,
-          title: "Retry transient gh failures",
-          body: "",
-          createdAt: "2026-03-13T00:00:00Z",
-          updatedAt: "2026-03-13T00:00:00Z",
-          url: "https://example.test/issues/105",
-          state: "OPEN",
-        },
-        {
-          issue_number: 105,
-          state: "draft_pr",
-          branch: "codex/issue-105",
-          pr_number: null,
-          workspace: "/tmp/workspaces/issue-105",
-          journal_path: null,
-          review_wait_started_at: null,
-          review_wait_head_sha: null,
-          copilot_review_requested_observed_at: null,
-          copilot_review_requested_head_sha: null,
-          copilot_review_timed_out_at: null,
-          copilot_review_timeout_action: null,
-          copilot_review_timeout_reason: null,
-          codex_session_id: null,
-          local_review_head_sha: null,
-          local_review_blocker_summary: null,
-          local_review_summary_path: null,
-          local_review_run_at: null,
-          local_review_max_severity: null,
-          local_review_findings_count: 0,
-          local_review_root_cause_count: 0,
-          local_review_verified_max_severity: null,
-          local_review_verified_findings_count: 0,
-          local_review_recommendation: null,
-          local_review_degraded: false,
-          last_local_review_signature: null,
-          repeated_local_review_signature_count: 0,
-          external_review_head_sha: null,
-          external_review_misses_path: null,
-          external_review_matched_findings_count: 0,
-          external_review_near_match_findings_count: 0,
-          external_review_missed_findings_count: 0,
-          attempt_count: 1,
-          implementation_attempt_count: 1,
-          repair_attempt_count: 0,
-          timeout_retry_count: 0,
-          blocked_verification_retry_count: 0,
-          repeated_blocker_count: 0,
-          repeated_failure_signature_count: 0,
-          last_head_sha: "deadbeef",
-          last_codex_summary: null,
-          last_recovery_reason: null,
-          last_recovery_at: null,
-          last_error: null,
-          last_failure_kind: null,
-          last_failure_context: null,
-          last_blocker_signature: null,
-          last_failure_signature: null,
-          blocked_reason: null,
-          processed_review_thread_ids: [],
-          processed_review_thread_fingerprints: [],
-          updated_at: "2026-03-13T00:00:00Z",
-        },
-      ),
-    /No commits between main and codex\/issue-105/,
-  );
-
-  assert.equal(calls, 1);
 });
