+++ src/github.ts
@@ -9,61 +9,21 @@ import {
   ReviewThread,
   SupervisorConfig,
 } from "./types";
-import { CommandOptions, CommandResult, runCommand } from "./command";
+import { CommandOptions, runCommand } from "./command";
 import {
   classifyConfiguredBotTopLevelReviewStrength,
   hasActionableReviewText,
   isActionableTopLevelReview,
 } from "./external-review-signal-heuristics";
+import { GitHubTransport } from "./github-transport";
+import type { GitHubCommandRunner } from "./github-transport";
 import { parseJson, truncate } from "./utils";
 
-const TRANSIENT_GITHUB_RETRY_LIMIT = 2;
-const TRANSIENT_GITHUB_RETRY_BASE_DELAY_MS = 200;
+export { isTransientGitHubCommandFailure } from "./github-transport";
+export type { GitHubCommandRunner } from "./github-transport";
+
 const COPILOT_REVIEW_TRANSITION_CACHE_TTL_MS = 30_000;
 const COPILOT_REVIEW_CACHE_MAX_ENTRIES = 128;
-
-export type GitHubCommandRunner = (
-  command: string,
-  args: string[],
-  options?: CommandOptions,
-) => Promise<CommandResult>;
-
-function sleep(ms: number): Promise<void> {
-  return new Promise((resolve) => setTimeout(resolve, ms));
-}
-
-export function isTransientGitHubCommandFailure(message: string | null | undefined): boolean {
-  if (!message) {
-    return false;
-  }
-
-  const lower = message.toLowerCase();
-  const githubRelated =
-    lower.includes("api.github.com") ||
-    lower.includes("github.com") ||
-    lower.includes("graphql") ||
-    lower.includes("gh ");
-  const transientSignal =
-    lower.includes("connection reset by peer") ||
-    lower.includes("unexpected eof") ||
-    lower.includes("eof") ||
-    lower.includes("tls handshake timeout") ||
-    lower.includes("i/o timeout") ||
-    lower.includes("timeout awaiting response headers") ||
-    lower.includes("temporary failure in name resolution") ||
-    lower.includes("no such host") ||
-    lower.includes("connection refused") ||
-    lower.includes("network is unreachable") ||
-    lower.includes("server closed idle connection") ||
-    lower.includes("http2: client connection lost") ||
-    lower.includes("stream error") ||
-    lower.includes("internal server error") ||
-    lower.includes("bad gateway") ||
-    lower.includes("service unavailable") ||
-    lower.includes("gateway timeout");
-
-  return githubRelated && transientSignal;
-}
 
 interface PullRequestStatusCheckRollupResponse {
   statusCheckRollup?: Array<{
@@ -427,87 +387,19 @@ function normalizeRollupChecks(rollup: PullRequestStatusCheckRollupResponse | nu
   return checks;
 }
 
-function summarizeGhArgs(args: string[]): string {
-  const visibleArgs = args.slice(0, 2);
-  const omittedCount = Math.max(args.length - visibleArgs.length, 0);
-  const summary = ["gh", ...visibleArgs].join(" ");
-  return omittedCount > 0 ? `${summary} +${omittedCount} arg${omittedCount === 1 ? "" : "s"}` : summary;
-}
-
-function sanitizeGhCommandMessage(message: string, args: string[]): string {
-  const commandSummary = summarizeGhArgs(args);
-  return message
-    .split("\n")
-    .map((line) => {
-      if (line.startsWith("Command failed: gh ")) {
-        return `Command failed: ${commandSummary}`;
-      }
-
-      if (line.startsWith("Command timed out: gh ")) {
-        return `Command timed out: ${commandSummary}`;
-      }
-
-      const timedOutAfterMarker = ": gh ";
-      if (line.startsWith("Command timed out after ") && line.includes(timedOutAfterMarker)) {
-        return `${line.slice(0, line.indexOf(timedOutAfterMarker) + 2)}${commandSummary}`;
-      }
-
-      return line;
-    })
-    .join("\n");
-}
-
 export class GitHubClient {
   private readonly copilotReviewLifecycleCache = new Map<string, CachedCopilotReviewLifecycleEntry>();
+  private readonly transport: GitHubTransport;
 
   constructor(
     private readonly config: SupervisorConfig,
-    private readonly commandRunner: GitHubCommandRunner = runCommand,
-    private readonly delay: (ms: number) => Promise<void> = sleep,
+    commandRunner = runCommand,
+    delay: (ms: number) => Promise<void> = async (ms) => {
+      await new Promise((resolve) => setTimeout(resolve, ms));
+    },
     private readonly now: () => number = Date.now,
-  ) {}
-
-  private async runGhCommand(args: string[], options: CommandOptions = {}): Promise<CommandResult> {
-    let lastTransientMessage: string | null = null;
-    const commandSummary = summarizeGhArgs(args);
-
-    for (let attempt = 0; attempt <= TRANSIENT_GITHUB_RETRY_LIMIT; attempt += 1) {
-      try {
-        const result = await this.commandRunner("gh", args, options);
-        if (result.exitCode === 0 || !isTransientGitHubCommandFailure(`${result.stderr}\n${result.stdout}`)) {
-          return result;
-        }
-
-        lastTransientMessage = truncate(
-          [result.stderr.trim(), result.stdout.trim()].filter(Boolean).join("\n"),
-          500,
-        );
-      } catch (error) {
-        const message = error instanceof Error ? error.message : String(error);
-        if (!isTransientGitHubCommandFailure(message)) {
-          throw error;
-        }
-
-        lastTransientMessage = truncate(sanitizeGhCommandMessage(message, args), 500);
-      }
-
-      const nextAttempt = attempt + 1;
-      if (nextAttempt > TRANSIENT_GITHUB_RETRY_LIMIT) {
-        break;
-      }
-
-      console.warn(`Transient GitHub CLI failure for ${commandSummary}; retry ${nextAttempt}/${TRANSIENT_GITHUB_RETRY_LIMIT}.`);
-      await this.delay(TRANSIENT_GITHUB_RETRY_BASE_DELAY_MS * nextAttempt);
-    }
-
-    throw new Error(
-      [
-        `Transient GitHub CLI failure after ${TRANSIENT_GITHUB_RETRY_LIMIT + 1} attempts: ${commandSummary}`,
-        lastTransientMessage ?? "Unknown transient GitHub failure.",
-      ]
-        .filter(Boolean)
-        .join("\n"),
-    );
+  ) {
+    this.transport = new GitHubTransport(commandRunner, delay);
   }
 
   private repoOwnerAndName(): { owner: string; repo: string } {
@@ -517,6 +409,10 @@ export class GitHubClient {
     }
 
     return { owner, repo };
+  }
+
+  private async runGhCommand(args: string[], options: CommandOptions = {}) {
+    return this.transport.run(args, options);
   }
 
   async authStatus(): Promise<{ ok: boolean; message: string | null }> {
